refactor: useMessageScroller を作成し、メッセージのスクロールに関する処理を統合
